--- express-project-starter/.env.example
@@ -1,4 +1,0 @@
-DB_USERNAME=<<db_username>>
-DB_PASSWORD=<<db_password>>
-DB_DATABASE=<<db_database>>
-DB_HOST=localhost
