--- express-project-starter/.gitignore
@@ -1,3 +1,0 @@
-**/**/.DS_Store
-.env
-node_modules
