+++ package.json
@@ -10,6 +10,7 @@
   "dependencies": {
     "connect-session-sequelize": "^7.0.4",
     "cookie-parser": "~1.4.4",
+    "csurf": "^1.11.0",
     "debug": "~2.6.9",
     "express": "~4.16.1",
     "express-session": "^1.17.1",
