+++ views/layout.pug
@@ -3,6 +3,6 @@ html
   head
       block head
         title="Express Starter App"
-        link(rel='stylesheet' href='/stylesheets/reset.css')
+        link(rel='stylesheet' href='')
   body
     block content
