+++ views/log-in.pug
@@ -1,0 +1,11 @@
+extends layout.pug
+block content
+    form(action="/login" method="POST")
+        input(type="hidden" name="_csrf" value=token)
+        label(for="email")
+        input(type="text" name="email")
+        label(for="password")
+        input(type="password" name="password")
+        label(for="stayLoggedIn") Stay logged In
+        input(type="checkbox" name="stayLoggedIn")
+        button Log in
